// src/gen/api/types/ListFriends.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Friend } from './Friend.ts'

export type ListFriendsQueryParams = {
  /**
   * @type string | undefined
   */
  cursor?: string
  /**
   * @minLength 1
   * @maxLength 50
   * @default 20
   * @type integer | undefined
   */
  limit?: number
}

/**
 * @description Cursor-paginated accepted friends
 */
export type ListFriends200 = {
  /**
   * @type array
   */
  data: Friend[]
  /**
   * @type object
   */
  pagination: {
    /**
     * @type string,null
     */
    nextCursor: string | null
    /**
     * @type boolean
     */
    hasNext: boolean
    /**
     * @type number
     */
    limit: number
  }
}

/**
 * @description Validation failed
 */
export type ListFriends400 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Missing or invalid access token
 */
export type ListFriends401 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Internal server error
 */
export type ListFriends500 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

export type ListFriendsQueryResponse = ListFriends200

export type ListFriendsQuery = {
  Response: ListFriends200
  QueryParams: ListFriendsQueryParams
  Errors: ListFriends400 | ListFriends401 | ListFriends500
}